Refactor the filter to have None as valid value to select rather than toggling the filter icon.
