+++ src/redux/actionTypes.js
@@ -21,8 +21,6 @@ export const CLEAR_ERROR = 'CLEAR_ERROR';
 export const UPDATE_FILTER_TEXT = 'UPDATE_FILTER_TEXT';
 
 export const UPDATE_FILTER_SELECT = 'UPDATE_FILTER_SELECT';
-export const ENABLE_FILTER_SELECT = 'ENABLE_FILTER_SELECT';
-export const DISABLE_FILTER_SELECT = 'DISABLE_FILTER_SELECT';
 
 export const UPDATE_SORT_SELECT = 'UPDATE_SORT_SELECT';
 export const UPDATE_SORT_ORDER = 'UPDATE_SORT_ORDER';
